// front/src/api/models/fileSummaryDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DataShare API
 * API de transfert de fichiers DataShare (US01-US10)
 * OpenAPI spec version: 1.0.0
 */
import type { FileSummaryDtoStatus } from "./fileSummaryDtoStatus";

export interface FileSummaryDto {
  id: string;
  originalName: string;
  sizeBytes: number;
  mimeType: string;
  downloadToken: string;
  expiresAt: string;
  createdAt: string;
  passwordProtected: boolean;
  status: FileSummaryDtoStatus;
  tags: string[];
}